+++ interfaces/chat-repl/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brains/chat-repl
 
+## 0.2.0-alpha.51
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/plugins@0.2.0-alpha.51
+
 ## 0.2.0-alpha.50
 
 ### Patch Changes