+++ shared/test-utils/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @brains/test-utils
 
+## 0.2.0-alpha.51
+
 ## 0.2.0-alpha.50
 
 ## 0.2.0-alpha.49