+++ shell/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @brains/core
 
+## 0.2.0-alpha.51
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/utils@0.2.0-alpha.51
+  - @brains/plugins@0.2.0-alpha.51
+  - @brains/templates@0.2.0-alpha.51
+
 ## 0.2.0-alpha.50
 
 ### Patch Changes